requirements.txt
gpiozero==2.0.1
paho-mqtt
